+++ react-sentry/src/Sentry.spec.tsx
@@ -2,6 +2,8 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { Example } from './Sentry.stories';
 
+import * as utils from './utils';
+
 describe('Sentry', () => {
   beforeEach(() => {
     // @ts-expect-error - missing other keys
@@ -22,5 +24,15 @@ describe('Sentry', () => {
 
     // eslint-disable-next-line no-console
     expect(console.warn).not.toBeCalled();
+  });
+
+  it('should call init', () => {
+    const mock = jest.fn();
+    const spy = jest.spyOn(utils, 'initSentry').mockImplementation(() => mock);
+
+    render(<Example dsn="test" />);
+    render(<Example dsn="test" />);
+
+    expect(spy).toBeCalledTimes(2);
   });
 });